move application related file to a directory
