+++ frontend/src/App.jsx
@@ -6,7 +6,7 @@ import JobDetail from './components/JobDetail';
 import WishlistPage from './components/wishlist/Wishlist';
 import Resume from './components/UserInfo/Resume';
 import Register from './components/Register';
-import UserApplications from './components/UserApplications';
+import UserApplications from './components/Application/UserApplications';
 import "./App.css";
