+++ frontend/src/components/Application/UserApplications.css
